requirements.txt
python-dotenv>=1.0.0
Flask==3.0.0
Flask-CORS==4.0.0
Pillow>=10.4.0
pytesseract==0.3.10
Werkzeug==3.0.1
# อ่านบาร์โค้ดเลขที่เช็คจากรูป (ถ้าไม่ติดตั้ง ระบบยังใช้ OCR/MICR ได้ตามเดิม)
pyzbar>=0.1.9
# ใช้ AI (Gemini) ดูภาพเช็คดึงเลขที่เช็ค (ต้องตั้ง GEMINI_API_KEY)
google-generativeai>=0.8.0
# ส่งออก/โหลดจาก Excel
openpyxl>=3.1.0
# ส่งข้อมูลไป Google Apps Script (รองรับ redirect ของ Google)
requests>=2.31.0